Add connection to second client server and format
Also has code reminiscent of adding a CSS style sheet.
Adding an HTML, CSS, and potentially Javascript file will be difficult
and hard to look at due to styling, long HTML code, and broken
relative file links. The next thing to do will be to move the HTML
text embeddded in Arduino and move it over to an individual HTML file.
This will allow easier formating.